refactor(hooks): lift useInfiniteScroll to top-level hooks/
The IntersectionObserver hook is fully generic — no experts-specific
knowledge — and the upcoming programs listing refactor wants the same
sentinel-driven infinite scroll behavior. Move it from
`app/explore/experts/hooks/useInfiniteScroll.ts` to top-level
`hooks/useInfiniteScroll.ts` so both listings can import from the
same place.

The experts barrel `app/explore/experts/hooks/index.ts` now re-exports
from the new location, so consumers using
`import { useInfiniteScroll } from "./hooks"` keep working without
any churn at the call sites.
