+++ app/explore/experts/hooks/index.ts
@@ -1,7 +1,10 @@
 export { useConsultants } from "./useConsultants";
 export { useCuratedExperts } from "./useCuratedExperts";
 export { useExpertsFilters } from "./useExpertsFilters";
-export { useInfiniteScroll } from "./useInfiniteScroll";
+// Re-export so existing `import { useInfiniteScroll } from "./hooks"`
+// keeps working after the hook was lifted to the top-level `hooks/`
+// directory for shared use across the experts and programs listings.
+export { useInfiniteScroll } from "@/hooks/useInfiniteScroll";
 export {
   useExpertFilterChips,
   type ChipKey,
